Refactoring code; no more 50,000 line per-file render backends!

+++ include/SpecialK/performance/io_monitor.h
@@ -45,13 +45,13 @@ struct io_perf_t {
   ULONGLONG      dt           =     0ULL;
   IO_COUNTERS    last_counter = {      };
 
-  double         read_mb_sec   = 0.0;
-  double         write_mb_sec  = 0.0;
-  double         other_mb_sec  = 0.0;
+  long double    read_mb_sec   = 0.0L;
+  long double    write_mb_sec  = 0.0L;
+  long double    other_mb_sec  = 0.0L;
 
-  double         read_iop_sec  = 0.0;
-  double         write_iop_sec = 0.0;
-  double         other_iop_sec = 0.0;
+  long double    read_iop_sec  = 0.0L;
+  long double    write_iop_sec = 0.0L;
+  long double    other_iop_sec = 0.0L;
 };
 
 void SK_CountIO (io_perf_t& ioc, const double update = 0.25 / 1.0e-7);
